// src/schemas/schema.ts
const componentValue: APG.Product = Object.freeze({
	type: "product",
	components: [
		componentComponentKey,
		componentComponentSource,
		componentComponentValue,
	],
})

Object.freeze(componentValue.components)

const componentLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.component,
	value: componentValue,
})

// Coproduct (index 1)
const coproductValue: APG.Unit = Object.freeze({ type: "unit" })
const coproductLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.coproduct,
	value: coproductValue,
})

// Iri (index 2)
const iriValue: APG.Unit = Object.freeze({ type: "unit" })
const iriLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.iri,
	value: iriValue,
})

// Label (index 3)
const labelComponentKeyValue: APG.Iri = Object.freeze({ type: "iri" })
const labelComponentKey: APG.Component = Object.freeze({
	type: "component",
	key: ns.key,
	value: labelComponentKeyValue,
})

const labelComponentValue: APG.Component = Object.freeze({
	type: "component",
	key: ns.value,
	value: anyValue,
})

const labelValue: APG.Product = Object.freeze({
	type: "product",
	components: [labelComponentKey, labelComponentValue],
})

Object.freeze(labelValue.components)

const labelLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.label,
	value: labelValue,
})

// Literal (index 4)
const literalComponentDatatypeValue: APG.Iri = Object.freeze({ type: "iri" })
const literalComponentDatatype: APG.Component = Object.freeze({
	type: "component",
	key: ns.datatype,
	value: literalComponentDatatypeValue,
})

const literalValue: APG.Product = Object.freeze({
	type: "product",
	components: [literalComponentDatatype],
})

Object.freeze(literalValue.components)

const literalLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.literal,
	value: literalValue,
})
// ...
const optionValue: APG.Product = Object.freeze({
	type: "product",
	components: [optionComponentKey, optionComponentSource, optionComponentValue],
})

Object.freeze(optionValue.components)

const optionLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.option,
	value: optionValue,
})

// Product (index 6)
const productValue: APG.Unit = Object.freeze({ type: "unit" })
const productLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.product,
	value: productValue,
})

// Reference (index 7)
const referenceComponentValueValue: APG.Reference = Object.freeze({
	type: "reference",
	value: 3,
})
const referenceComponentValue: APG.Component = Object.freeze({
	type: "component",
	key: ns.value,
	value: referenceComponentValueValue,
})

const referenceValue: APG.Product = Object.freeze({
	type: "product",
	components: [referenceComponentValue],
})

Object.freeze(referenceValue.components)

const referenceLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.reference,
	value: referenceValue,
})

// Unit (index 8)
const unitValue: APG.Unit = Object.freeze({ type: "unit" })
const unitLabel: APG.Label = Object.freeze({
	type: "label",
	key: ns.unit,
	value: unitValue,
})

const schemaSchema: [
	typeof componentLabel,
	typeof coproductLabel,
	typeof iriLabel,
	typeof labelLabel,
	typeof literalLabel,
	typeof optionLabel,
	typeof productLabel,
	typeof referenceLabel,
	typeof unitLabel
] = [
	componentLabel,
	coproductLabel,
	iriLabel,
	labelLabel,
	literalLabel,
	optionLabel,
	productLabel,
	referenceLabel,
	unitLabel,
]